Changed output of i18n block expression to output span element
Now each translation will be wrapped in an span element with two
attributes, "data-i18n-namespace" and "data-i18n-key", which identifies
the translation in the DOM. Furthermore, if the expression contains a
parameter with the placeholder "attribute:", the expression outputs an
attribute. E.g. {{i18n "attributes:placeholder" "Test"}} ->
placeholder="Test".